Add operator Precedence
Expressions are first parsed into values and operators.
Then each operator type is checked in order.
The values on each side of the operator are merged as a new expression
This is repeated until there is only a single expression.

This also caught a few bugs in my test cases wherein expressions had
dangling operators or values next to one another.